$or array support, $upsert support, $merge support
Fixed the broken logic for the $or array.
Created tests to validate.

Modified save logic to only create the collection on $upsert.
Created tests to validate.

Allow an options object to be passed into the EJDBDatabase.save() so you can request that the objects be merged using the $merge key.
Created tests to validate.